Added efficient_kan and refactored main transformer model code into 1 file for BERT and GPT
