--- models/NLPModels/nanoGPT/.gitattributes
@@ -1,3 +1,0 @@
-# Override jupyter in Github language stats for more accurate estimate of repo code languages
-# reference: https://github.com/github/linguist/blob/master/docs/overrides.md#generated-code
-*.ipynb linguist-generated
