--- models/NLPModels/nanoGPT/.gitignore
@@ -1,12 +1,0 @@
-.DS_Store
-.idea
-.ipynb_checkpoints/
-.vscode
-__pycache__/
-*.bin
-*.pkl
-*.pt
-*.pyc
-input.txt
-env/
-venv/
