+++ models/NLPModels/nanos/.gitignore
@@ -10,3 +10,4 @@ __pycache__/
 input.txt
 env/
 venv/
+*.egg-info/
